feat(collect-modal): component separation
